+++ src/client/java/com/rubyimpala/features/auction/models/PriceEntry.java
@@ -4,12 +4,12 @@ import java.util.Map;
 
 public record PriceEntry(
         String id,
-        int totalPrice,
+        long totalPrice,
         int count,
         long timestamp,
         Map<String, Integer> enchants
 ) {
-    public int getUnitPrice(){return totalPrice() / count();}
+    public long getUnitPrice(){return totalPrice() / count();}
 //    public int getStackPrice() {return price * 64;}
 }
